+++ changelog.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-Last change: 01-MAR-2026 CEST
+Last change: 02-MAR-2026 CEST
 
 0.9.8
 -----
@@ -10,7 +10,7 @@ Last change: 01-MAR-2026 CEST
 - Localize Parameters dialog displays a message if the z calibration path in the config file could not be found
 - Fixed Picasso Server launching in one-click-installers
 - ``picasso.localize.fit`` default method changed to ``sigmaxy`` (anisotropic sigma fitting)
-.. - Fixed 2D MLE fitting and cleaned the docstrings for better readability (``picasso.gaussmle``)
+- Fixed 2D MLE fitting and cleaned the docstrings for better readability (``picasso.gaussmle``)
 
 0.9.7
 -----